worker/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
redis>=5.0.0
